fix(node-amqp-connection-manager): forces back a dependency that has a bug.
node-amqp-connection-manager, starting with v3.4.0, has a bug where connectionOptions are not passed into the underlying AMQPLIB. This caused AMQPCaccon to not be able to connect to self-signed TLS servers, for example, nor any other server that requires connectionOptions.

For now, package.json of this library is keeping node-amqp-connection-manager at 3.3.x using the syntax "~3.3.0" since the old syntax "^3.3.2" would in fact allow the dependency to go up to 3.x (and the bug is present in 3.4 and up.)

This might be reverted in a future version.